+++ CHANGES.rst
@@ -4,6 +4,14 @@ Changelog
 3.2.4 (unreleased)
 ------------------
 
+Features:
+
+- Define ``download`` and ``display-file`` views that work for tiles.
+  The original views in ``plone.namedfile`` cannot find the tile data.
+  [maurits]
+
+Bug fixes:
+
 - Fixed getting original image from tile.
   Until now, the ``images`` view tried to get the field from the tile instead of the tile data.
   This only worked when you had explicitly added a property with this field name on the tile.